requirements.txt
# Base
python-dotenv
jinja2
pyyaml
ujson
uvloop
asyncio
aiohttp
aioredis
aiojobs
jsonrpcserver 
git+https://github.com/bcb/jsonrpcclient.git@master#egg=jsonrpcclient
docker
aiofiles
aiojobs
# bots
aiogram
